services/event-mapper/requirements.txt
aiokafka==0.10.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
structlog==23.2.0
python-dotenv==1.0.0
